% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{abp}
\alias{abp}
\title{Simple 125 hz Arterial blood pressure waveform}
\format{
An object of class \code{numeric} of length 5000.
}
\usage{
data(abp)
}
\description{
Simple 125 hz Arterial blood pressure waveform
}
\keyword{datasets}
