% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventilator_waveform_processing.R
\name{add_volume_signal}
\alias{add_volume_signal}
\title{Generate volume waveform from flow}
\usage{
add_volume_signal(vent_df)
}
\arguments{
\item{vent_df}{ventilation waveform data frame}
}
\description{
Generate volume waveform from flow
}
