% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{anonymize_record_time}
\alias{anonymize_record_time}
\title{Anonymize record time}
\usage{
anonymize_record_time(rec)
}
\arguments{
\item{rec}{(nested) list of signals}
}
\description{
Shift all time variables by a random amount between 0 and ~32 years (1e9 secs)
}
