% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_signal.R
\name{dygraph_events}
\alias{dygraph_events}
\title{Add events to dygraph}
\usage{
dygraph_events(dygraph, events, label_col = 2, time_col = 1)
}
\arguments{
\item{dygraph}{Dygraph to add events to}

\item{events}{Events dataframe}

\item{label_col}{Column containing labels}

\item{time_col}{Column to use for x-axis}
}
\description{
Add events to dygraph
}
