% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_tools.R
\name{dygraph_gam}
\alias{dygraph_gam}
\title{Plot model predictions and observations using Dygraphs}
\usage{
dygraph_gam(..., resid = FALSE)
}
\arguments{
\item{...}{One or more Gam models.}

\item{resid}{include residuals}
}
\description{
Plot model predictions and observations using Dygraphs
}
