% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_signal.R
\name{dygraph_signal}
\alias{dygraph_signal}
\title{Plot a signal with dygraphs}
\usage{
dygraph_signal(
  signal,
  ...,
  time_col = 1,
  main = NULL,
  height = NULL,
  width = "100\%",
  maxlen = 2e+05
)
}
\arguments{
\item{signal}{Signal dataframe}

\item{...}{Tidyselect argument to select signal columns to plot. If empty, plot everything.}

\item{time_col}{Column to use for x-axis}

\item{main}{Graph title}

\item{height}{height of dygraph HTML element}

\item{width}{Width of dygraph HTML element}

\item{maxlen}{maximum number of samples.
The signal is down sampled to this number.}
}
\description{
Plot a signal with dygraphs
}
