% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{ecg_noisy}
\alias{ecg_noisy}
\title{Noisy ECG (500 hz)}
\format{
An object of class \code{numeric} of length 9001.
}
\usage{
data(ecg_noisy)
}
\description{
Noisy ECG (500 hz)
}
\keyword{datasets}
