% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_waveforms.R
\name{find_abp_beats}
\alias{find_abp_beats}
\title{Create DataFrame of arterial pressure waveform annotations.}
\usage{
find_abp_beats(
  data,
  abp_col = 2,
  time_col = 1,
  min_PP = 0.2,
  min_beat_width_s = 0.3,
  win_size_avg = 2000,
  show.plot = FALSE,
  include_waveform = FALSE,
  sample_rate = NULL
)
}
\arguments{
\item{data}{Vector of arterial blood pressure.}

\item{abp_col}{Index or name of column with abp data}

\item{time_col}{Vector with the same length as \code{abp}, used to indicate the timing of each sample.
If this variable is provided, sample indexes will be substituted for this vector.}

\item{min_PP}{Minimum ratio of PP/mean pressure.
If a beat has a lower PP, it is considered a fluctuation around the threshold.
The dicrotic notch is a common cause of such fluctuations.}

\item{min_beat_width_s}{Minimum beat width (in seconds)
The default is 0.3 seconds}

\item{win_size_avg}{Window size for moving avg calculation.
This is used to calculate the cutoff value for when a peak represents a new beat and not just noise.
Lower values increase flexibility. Use visualize_abp_peak_detection() to dial in.}

\item{show.plot}{Show diagnostic plot}

\item{include_waveform}{Include waveform for each beat in output}

\item{sample_rate}{Used if data is a vector of samples.}
}
\description{
Position and value of systolic and diastolic pressures.
Each row represents one beat with first a diastolic and then a systolic value.
}
\details{
.noise_pos_after_sys is the sum of positive changes in ABP excluding the
systole and the dicrotic notch, divided by the beat length in seconds.
2 mmhg/s seems like a good cutoff.
}
