% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_waveforms.R
\name{fix_repeating_values}
\alias{fix_repeating_values}
\title{Remove repeating instances of the same sample}
\usage{
fix_repeating_values(df, expected_interval = 20, max_difftime = 25)
}
\arguments{
\item{df}{Signal to fix (data frame)}

\item{expected_interval}{The interval expected from the device.}

\item{max_difftime}{If the interval between two changes is higher than max_timediff,
this is interpreted as representing multiple true measurements.}
}
\value{
Dataframe with only relevant values (1 every 20 seconds)
}
\description{
A device may only return a value every 20 seconds. If data is represented with 2 sec intervals, 9/10 are just
repeats. This function takes a dataframe with repeating values, and tries to only return 1 observation per
measured value.
}
\details{
eg. Vital recorder saves EV1000 samples every other second. Data is only calculated every 20 seconds.
This function tries to match these 20 second samples.
}
