% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_waveforms.R
\name{fix_time_shifts}
\alias{fix_time_shifts}
\title{Fix Time shifts in vital waveforms}
\usage{
fix_time_shifts(
  signal_list,
  skip_shift = NULL,
  show_diagnostics = TRUE,
  diagnostic_signal = "CVP"
)
}
\arguments{
\item{signal_list}{list of vital signals}

\item{skip_shift}{Index (vector) of shift to skip (allow in output)}

\item{show_diagnostics}{Show diagnostic prints and plots}

\item{diagnostic_signal}{Signal to use in diagnostic plots}
}
\description{
Vital Recorder may introduce short shifts in time in waveform recordings.
This is probably an error, and can be removed.
This function looks through a waveform and for every diff higher than 1/samplerate,
subtracts this diff - samplerate from all signal-times that are later than the culprit.
}
