% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_waveforms.R
\name{flag_beats}
\alias{flag_beats}
\title{Flag Abnormal Beats}
\usage{
flag_beats(beats_df, max_pos_after_sys = 2, max_PP_change = 0.15)
}
\arguments{
\item{beats_df}{Dataframe of beats (from \code{find_abp_beats()})}

\item{max_pos_after_sys}{Maximum allow positive deflection excluding 2 (systole and dicrotic notch)
mmHg/s}

\item{max_PP_change}{Maximum allowed PP change (\%) compared to median.}
}
\value{
a vector indicating whether a beat is noisy / abnormal
}
\description{
Flag Abnormal Beats
}
