% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{format_hms}
\alias{format_hms}
\title{Format as HMS with 3 decimals}
\usage{
format_hms(time_stamp)
}
\arguments{
\item{time_stamp}{Timestamp to format}
}
\description{
Format as HMS with 3 decimals
}
