% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_slow_data.R
\name{gen_device_settings_annotation}
\alias{gen_device_settings_annotation}
\title{Generate simple data frame of device settings}
\usage{
gen_device_settings_annotation(dev_settings, min_setting_len = 0)
}
\arguments{
\item{dev_settings}{Data frame of device settings, from vscapture}

\item{min_setting_len}{Settings must be changed one at a time.
if a setting is kept for shorter than this duration (in seconds) it is considered
to be part of a change.}
}
\description{
Generate simple data frame of device settings
}
