% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_waveforms.R
\name{gen_r_peaks_df}
\alias{gen_r_peaks_df}
\title{Find R peaks in ECG}
\usage{
gen_r_peaks_df(ecg_df, freq = NULL, ...)
}
\arguments{
\item{ecg_df}{Vector of ecg waveform with a}

\item{freq}{Sample frequency}

\item{...}{Passed to detect_rpeaks()}
}
\description{
Wrapper for detect_rpeaks() (Pan-Tompkins algorithm from {rsleep}).
}
