% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal.R
\name{geom_annotation}
\alias{geom_annotation}
\title{Geom to add point annotations to plot}
\usage{
geom_annotation(signal, seconds = NULL)
}
\arguments{
\item{signal}{must be an EVENT signal}

\item{seconds}{Limit plot to first x seconds}
}
\description{
Geom to add point annotations to plot
}
