% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_tools.R
\name{get_gam_ppv}
\alias{get_gam_ppv}
\title{Get PPV from GAM model}
\usage{
get_gam_ppv(model, term, ppv_only = FALSE)
}
\arguments{
\item{model}{Gam model}

\item{term}{term to calculate ppv from (e.g. 's(resp_index)')}

\item{ppv_only}{Return only PPV value. Otherwise return row of related parameters as well.}
}
\description{
Get PPV from GAM model
}
