% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_tools.R
\name{get_PP_gam_predictions}
\alias{get_PP_gam_predictions}
\title{Create data frame of data, fit and residuals of pulse pressure GAM}
\usage{
get_PP_gam_predictions(PP_gam)
}
\arguments{
\item{PP_gam}{GAM of PP}
}
\value{
data frame
}
\description{
Create data frame of data, fit and residuals of pulse pressure GAM
}
