% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_waveforms.R
\name{is_extra_systole}
\alias{is_extra_systole}
\title{Find extra systoles from beat times}
\usage{
is_extra_systole(beat_times, threshold = 0.1)
}
\arguments{
\item{beat_times}{time of each beat (vector of datetimes)}

\item{threshold}{if a beat interval (RR) is shorter than the median of
the preceding 10 intervals by more than this threshold (relative), it is
considered and extra systole (default = 0.1, 10\%).}
}
\description{
Find extra systoles from beat times
}
