% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{join_nearest}
\alias{join_nearest}
\title{Join Nearest
Left join two data frames, by aligning the nearest xkey with the nearest ykey.}
\usage{
join_nearest(x, y, xkey = "time", ykey = "time")
}
\arguments{
\item{x}{left data frame}

\item{y}{right data frame}

\item{xkey, ykey}{keys for x and y, Defaults to joining by time}
}
\description{
Join Nearest
Left join two data frames, by aligning the nearest xkey with the nearest ykey.
}
