% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_tools.R
\name{plot_PP_gam}
\alias{plot_PP_gam}
\title{Plot pulse pressure GAM smooths and partial residuals}
\usage{
plot_PP_gam(PP_gam, return_list = FALSE)
}
\arguments{
\item{PP_gam}{PP GAM}

\item{return_list}{If true, return list of plots instead of combined plot.}
}
\value{
Patchworked ggplot
}
\description{
Plot pulse pressure GAM smooths and partial residuals
}
