% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot a signal}
\usage{
plot_signal(signal, signal_col = 2, seconds = NULL)
}
\arguments{
\item{signal}{Waveform data frame with time and value}

\item{signal_col}{Column with y values}

\item{seconds}{Limit plot to first x seconds}
}
\description{
Plot a signal
}
