% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{PP_gam}
\alias{PP_gam}
\title{mgcv GAM of pulse pressures}
\format{
An object of class \code{gam} (inherits from \code{glm}, \code{lm}) of length 53.
}
\usage{
data(PP_gam)
}
\description{
mgcv GAM of pulse pressures
}
\keyword{datasets}
