% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vital_csv.R
\name{read_vital}
\alias{read_vital}
\title{Read Vital}
\usage{
read_vital(
  path,
  tz = "UTC",
  nested_list = TRUE,
  tracks_only = FALSE,
  recalc_time = FALSE
)
}
\arguments{
\item{path}{Path to folder containing the files exported by parse_vital}

\item{tz}{Time zone}

\item{nested_list}{Create a nested list of tracks inside a list of devices.
Necessary to deal with duplicate track names between devices.}

\item{tracks_only}{Do not include header in returned list.
If FALSE, an error is given if there are any duplicate track names.}

\item{recalc_time}{recalculate time using first value and sample rate.
(Should probably not be used).}
}
\value{
A (nested) list of tracks.
}
\description{
Loads a folder of gz compressed csv files, exported by parse_vital
}
\examples{
\dontrun{
read_vital(folder, tz = 'CET')
}
}
