% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vital_csv.R
\name{read_vital_csv}
\alias{read_vital_csv}
\title{Load csv files created by custom python module.}
\usage{
read_vital_csv(path, var_name = "value", tz = "UTC")
}
\arguments{
\item{path}{Path to .csv file}

\item{var_name}{name given to the values column (e.g. CVP)}

\item{tz}{time zone}
}
\description{
Load csv files created by custom python module.
}
