% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vital_csv.R
\name{read_vital_header}
\alias{read_vital_header}
\title{Read Header CSV (tracks.csv) from parse_vital}
\usage{
read_vital_header(path)
}
\arguments{
\item{path}{Path to folder containing the files exported by parse_vital}
}
\description{
Read Header CSV (tracks.csv) from parse_vital
}
