% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vscapture.R
\name{read_vscapture}
\alias{read_vscapture}
\title{Load all csv files created by VSCapture DrgVent.
Converts column types}
\usage{
read_vscapture(path, start = NULL, end = NULL, show_diagnostics = TRUE)
}
\arguments{
\item{path}{Path to folder containing vital capture .csv files}

\item{start, end}{start and end time for import}

\item{show_diagnostics}{show plot of expected vs recorded times (recorded times are wrong, but should be correct on average).}
}
\description{
Load all csv files created by VSCapture DrgVent.
Converts column types
}
