% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vscapture.R
\name{read_vscapture_rt}
\alias{read_vscapture_rt}
\title{Load csv files with realtime data created by VSCapture DrgVent}
\usage{
read_vscapture_rt(
  path,
  tz = "CET",
  return_raw = FALSE,
  start = NULL,
  end = NULL,
  show_diagnostics = TRUE
)
}
\arguments{
\item{path}{Path to VSCapture csv folder}

\item{tz}{time zone}

\item{return_raw}{return original flow and pressure tables and new corrected times as separate entries in a list.}

\item{start, end}{start and end time for import}

\item{show_diagnostics}{show plot of expected vs recorded times (recorded times are wrong, but should be correct on average).}
}
\description{
Load csv files with realtime data created by VSCapture DrgVent
}
