% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vscapture.R
\name{read_vscapture_rt_csv}
\alias{read_vscapture_rt_csv}
\title{Load csv file with realtime data created by VSCapture DrgVent}
\usage{
read_vscapture_rt_csv(path)
}
\arguments{
\item{path}{Path to .csv file}
}
\description{
Load csv file with realtime data created by VSCapture DrgVent
}
