% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vscapture.R
\name{read_vscapture_slow}
\alias{read_vscapture_slow}
\title{Load all csv files with slow data created by VSCapture DrgVent}
\usage{
read_vscapture_slow(path, start = NULL, end = NULL)
}
\arguments{
\item{path}{Path to folder containing vital capture .csv files}

\item{start, end}{start and end time for import}
}
\description{
Load all csv files with slow data created by VSCapture DrgVent
}
