% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vscapture.R
\name{read_vscapture_slow_csv}
\alias{read_vscapture_slow_csv}
\title{Load csv files with slow data created by VSCapture DrgVent}
\usage{
read_vscapture_slow_csv(path, start = NULL, end = NULL)
}
\arguments{
\item{path}{Path to .csv file}

\item{start, end}{start and end time for import}
}
\description{
Load csv files with slow data created by VSCapture DrgVent
}
