% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_record}
\alias{sample_record}
\title{Complete recording (old format)}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(sample_record)
}
\description{
Complete recording (old format)
}
\keyword{datasets}
