% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_record2}
\alias{sample_record2}
\title{Complete recording (new format)}
\format{
An object of class \code{list} of length 10.
}
\usage{
data(sample_record2)
}
\description{
Complete recording (new format)
}
\keyword{datasets}
