% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_vscapture_rt}
\alias{sample_vscapture_rt}
\title{Waveform data from VS capture}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(sample_vscapture_rt)
}
\description{
Waveform data from VS capture
}
\keyword{datasets}
