% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_vscapture_slow}
\alias{sample_vscapture_slow}
\title{Slow data from VS capture}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(sample_vscapture_slow)
}
\description{
Slow data from VS capture
}
\keyword{datasets}
