% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{seconds_since_start}
\alias{seconds_since_start}
\title{Return seconds since start of time vector}
\usage{
seconds_since_start(x)
}
\arguments{
\item{x}{time vector}
}
\description{
Return seconds since start of time vector
}
