% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{shift_record_time}
\alias{shift_record_time}
\title{Shift record time}
\usage{
shift_record_time(rec, seconds)
}
\arguments{
\item{rec}{(nested) list of signals}

\item{seconds}{Time interval to shift by}
}
\description{
Shift all time variables
}
