% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{splitAt}
\alias{splitAt}
\title{Split by index}
\usage{
splitAt(x, pos, trim_ends = FALSE)
}
\arguments{
\item{x}{input vector}

\item{pos}{index vector}

\item{trim_ends}{exclude first and last group}
}
\description{
Similar to \code{split()} but takes a vector of indexes instead of a group vector.
From SO: \url{https://stackoverflow.com/a/19274414/1498656}
}
