% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{subset_record}
\alias{subset_record}
\title{Subset nested signals}
\usage{
subset_record(
  rec,
  starttime,
  endtime = NULL,
  relative_time = TRUE,
  absolute_time = TRUE
)
}
\arguments{
\item{rec}{(nested) list of signals}

\item{starttime}{start of interval (or interval)}

\item{endtime}{end of interval}

\item{relative_time}{include time since subset start as time_s}

\item{absolute_time}{include original absolute time.
If absolute time is not included, relative time overwrites the time column}
}
\description{
Subset a record of nested signals to a specific interval
}
